Added a new 'Label Attributes' (i.e. SelectAttributes) UI widget that allows a user to simultaneously specify a smaller subregion for each attribute value and label it at the same time, in a manner similar to LabelDigits.
Squashed commit of the following:

commit 71a60b4a0dac0b37f9790f7e5335a51825b2efa6
Author: Eric Kim <erickim555@gmail.com>
Date:   Mon Oct 22 13:15:50 2012 -0700

    SelectAttributes: Added a way to hide/show labeled attribute patches, for
    ease of use.

commit 7610a504c53913dd5a0a0f4e91e3e1c13b611d4d
Author: Eric Kim <erickim555@gmail.com>
Date:   Mon Oct 22 13:00:42 2012 -0700

    SelectAttributes: Don't allow mouse events to occur outside of the
    bitmap.

commit f8a7a76b8bd04591a9930f10498d7314ff9baa1c
Author: Eric Kim <erickim555@gmail.com>
Date:   Mon Oct 22 12:38:11 2012 -0700

    SelectAttributes: Increased font size for displayed attr name. Disabled
    creating new boxes on patches that already have a box.

commit 70482bae103293fdfd4e7a9d044e97501a6abd0a
Author: Eric Kim <erickim555@gmail.com>
Date:   Mon Oct 22 12:25:05 2012 -0700

    SelectAttributes: Made the data entry more convenient by processing
    text enter events.

commit 04ff61e89822420c4e85eeeb1388425e88e7ecc1
Author: Eric Kim <erickim555@gmail.com>
Date:   Mon Oct 22 12:21:54 2012 -0700

    SelectAttributes: Added a way to adjust the tempmatch threshold.

commit dde8c81c39dfa4eda0b40f56bcb31d7afbb72f02
Author: Eric Kim <erickim555@gmail.com>
Date:   Mon Oct 22 11:55:41 2012 -0700

    Minor typo.

commit 25e8201d251c48f28d9ebba389a3424fa1877d23
Author: Eric Kim <erickim555@gmail.com>
Date:   Mon Oct 22 11:51:00 2012 -0700

    Bug fixes: Sped up a pre-processing step in grouping. Fixed a bug in
    multiple-exemplar creation (where, for digit mult-exemplars, the wrong
    region was being used, i.e. a too-large region). Added debugging code.

commit 781dadb8e8d5c2205cfb4885b9b0d218ec481653
Author: Eric Kim <erickim555@gmail.com>
Date:   Sun Oct 21 17:14:18 2012 -0700

    SelectAttributes: More OpenCount+grouping integration. Introduced a
    TM_THRESHOLD that only accepts matches above a certain threshold.

commit 9a3b1bd8af3ef1c2fafb4af7072026cd739d2fd4
Author: Eric Kim <erickim555@gmail.com>
Date:   Sat Oct 20 23:40:59 2012 -0700

    Don't include digit-attributes in SelectAttributes.

commit 7f55a38313b36458866b43c500d295aea1ed105a
Author: Eric Kim <erickim555@gmail.com>
Date:   Sat Oct 20 17:52:29 2012 -0700

    SelectAttributes: Fixed a few OpenCount-integration bugs (i.e. use
    subpatches rather than the large patches). Hooked in the mult-exemplar
    creation code. Grouping seems to be successful now.

commit ca9bb4f1d4eb27120482df7f95e4d9b06dafc8c2
Author: Eric Kim <erickim555@gmail.com>
Date:   Sat Oct 20 16:46:36 2012 -0700

    SelectAttributes: More-fully hooked in the new UI into OpenCount.
    Added exporting of results, save/restore session state, and fixed
    a bug where clicking on the dummy cells would trigger an error.

commit bb127a6db159cf6093ed0025835ce8d84b650598
Author: Eric Kim <erickim555@gmail.com>
Date:   Sat Oct 20 15:46:41 2012 -0700

    SelectAttributes: Fixed UI bugs that resulted from hooking in the new UI
    into OpenCount. Added the next/prev attribute buttons.

commit 9a1ba2c15cb0741bf8fc2fda1c4ac8eb76c32e0c
Author: Eric Kim <erickim555@gmail.com>
Date:   Thu Oct 18 17:07:29 2012 -0700

    Initial work to integrate the new SelectAttributes UI into OpenCount.

commit 2a1e253cc863678ff4bea8428162f2fa3349e6f3
Author: Eric Kim <erickim555@gmail.com>
Date:   Thu Oct 18 16:37:05 2012 -0700

    Only run tempmatching on unlabeled images. Added a PREVMATCHES option
    to tempmatch.py.

commit 86cdc9f96d5f2b781d4ef8ef5ed71644d674c1af
Author: Eric Kim <erickim555@gmail.com>
Date:   Thu Oct 18 16:19:33 2012 -0700

    Hooked in the template matching results so that the results are stored
    and visualized.

commit 20e9cde6caf10b45601fac7c441a5e917a9f31cd
Author: Eric Kim <erickim555@gmail.com>
Date:   Thu Oct 18 15:37:14 2012 -0700

    Make the user label what the attribute value is before doing the
    tempmatching.

commit 1d5d59389d002940af4edfee10def69970713c01
Author: Eric Kim <erickim555@gmail.com>
Date:   Tue Oct 16 22:44:50 2012 -0700

    More progress on new select attributes UI. Hooked in a rudimentary
    tempmatch+verify-matching ui.

commit 46679524a6dbcc78c64ce2dee96e2a74858b5878
Author: Eric Kim <erickim555@gmail.com>
Date:   Tue Oct 16 20:11:26 2012 -0700

    Committed a tempmatch.py module, which consolidates all the various
    template matching I've been doing, and also providing a parallellized
    interface.
    First steps towards integration of template matching to new select
    attributes ui.

commit 8b0c4ea3da3688319b3104a5eec549e80d6fcba8
Author: Eric Kim <erickim555@gmail.com>
Date:   Tue Oct 16 18:42:57 2012 -0700

    More progress on the new define ballot attributes ui. Currently have
    the UI at a point where the user can draw a box on every cell. Next
    is to do the template matching stuff.

commit 231ed958aec5b7573372126fe17bee05c05d963f
Author: Eric Kim <erickim555@gmail.com>
Date:   Tue Oct 16 17:15:48 2012 -0700

    Allow you to specify numrows/cols/cellheight.

commit c7aa147a4359861a1263017e76a10960f9e0badb
Author: Eric Kim <erickim555@gmail.com>
Date:   Tue Oct 16 17:11:42 2012 -0700

    Initial commit of skeleton new UI for selecting attribute patches.
    Also fixed a few UI issues in MosaicPanel.

+++ opencount/digits_ui/digits_ui.py
@@ -84,13 +84,14 @@ class LabelDigitsPanel(wx.lib.scrolledpanel.ScrolledPanel):
                                                   project.extracted_digitpatch_dir)
         digit_ex_fulldir = pathjoin(project.projdir_path, project.digit_exemplars_outdir)
         precinctnums_fullpath = pathjoin(project.projdir_path, project.precinctnums_outpath)
-        _t = time.time()
-        print "Extracting Digit Patches..."
-        patch2temp = do_extract_digitbased_patches(self.project)
-        print "...Finished Extracting Digit Patches ({0} s).".format(time.time() - _t)
-        pickle.dump(patch2temp, open(pathjoin(project.projdir_path,
-                                              project.digitpatch2temp),
-                                     'wb'))
+        if not os.path.exists(extracted_digitpatches_fulldir):
+            print "Extracting Digit Patches..."
+            _t = time.time()
+            patch2temp = do_extract_digitbased_patches(self.project)
+            print "...Finished Extracting Digit Patches ({0} s).".format(time.time() - _t)
+            pickle.dump(patch2temp, open(pathjoin(project.projdir_path,
+                                                  project.digitpatch2temp),
+                                         'wb'))
         self.mainpanel = DigitMainPanel(self, extracted_digitpatches_fulldir,
                                         digit_ex_fulldir,
                                         precinctnums_fullpath,
@@ -306,7 +307,7 @@ class DigitLabelPanel(wx.lib.scrolledpanel.ScrolledPanel):
         for regionpath, digits_str in digits.iteritems():
             i, j = self.imgID2cell[regionpath]
             k = (self.NUM_COLS * i) + j
-            self.precinct_txts[regionpath].SetLabel("{0}: Precinct Number: {1}".format(str(k),
+            self.precinct_txts[regionpath].SetLabel("{0} Precinct Number: {1}".format(str(k),
                                                                                        digits_str))
         #for regionpath, boxes in cell_boxes.iteritems():
         #    self.cells[regionpath].boxes = boxes
@@ -673,7 +674,7 @@ digit.")
         for regionpath, digits_str in digits.iteritems():
             i, j = self.imgID2cell[regionpath]
             k = (self.NUM_COLS * i) + j
-            self.precinct_txts[regionpath].SetLabel("{0}: Precinct Number: {1}".format(str(k),
+            self.precinct_txts[regionpath].SetLabel("{0} Precinct Number: {1}".format(str(k),
                                                                                        digits_str))
         
         self.gridsizer.Layout()
@@ -760,7 +761,8 @@ are not of proper length.'
             assert patchpath not in result
             # txt.GetLabel() returns something like 'Precinct Number:0013038',
             # so get rid of this.
-            result[patchpath] = txt.GetLabel().split(":")[1].strip()
+            digitstr = txt.GetLabel().split(":")[1].strip()
+            result[patchpath] = digitstr
         return result
 
     def export_precinct_nums(self, result):
